refactor: 复用 lievisual 的 Color 与文本排版类型
- 用 lievisual::Color 替换 crate::color::Color
- 用 lievisual::text 的类型替换本地 FontWeight/FontStyle/TextAlign/TextDecoration
- 将 Color::new 迁移为 Color::rgb
- 更新 lievisual 依赖为本地路径，并同步 Cargo.lock

+++ src/output/svg.rs
@@ -12,7 +12,7 @@ pub fn document_to_svg(
 ) -> String {
     let scene = crate::document::to_scene::document_to_scene(document, settings, 72.0);
 
-    let mut r = lievisual::render::SvgRenderer::new(scene.width as f64, scene.height as f64)
+    let mut r = lievisual::render::SvgRenderer::new(scene.width, scene.height)
         .with_background(lievisual::geometry::Color::WHITE);
     use lievisual::render::Renderer;
     r.render_scene(&scene);
